6theng-assignement/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pypdf
tiktoken
streamlit
python-dotenv
